refactor dns_sd.h code in preparation for new mdnsd
