+++ lib/dns_sd/dns_sd.c
@@ -1,0 +1,404 @@
+/**
+ *  Copyright (C) 2011-2012  Juho Vähä-Herttua
+ *
+ *  This library is free software; you can redistribute it and/or
+ *  modify it under the terms of the GNU Lesser General Public
+ *  License as published by the Free Software Foundation; either
+ *  version 2.1 of the License, or (at your option) any later version.
+ *
+ *  This library is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *  Lesser General Public License for more details.
+ *
+ *=================================================================
+ * modified by fduncanh 2022, 2026
+ */
+
+#include <stdlib.h>
+#include <string.h>
+#include <stdio.h>
+#include <assert.h>
+
+#include "../compat.h"
+#include <dns_sd.h>
+#include "../dnssd.h"
+#include "../dnssdint.h"
+#include "../utils.h"
+
+#define MAX_DEVICEID 18
+#define MAX_SERVNAME 256
+
+#if defined(HAVE_LIBDL) && !defined(__APPLE__)
+# define USE_LIBDL 1
+#else
+# define USE_LIBDL 0
+#endif
+
+#if defined(_WIN32) || USE_LIBDL
+# ifdef _WIN32
+#  include <stdint.h>
+#  if !defined(EFI32) && !defined(EFI64)
+#   define DNSSD_STDCALL __stdcall
+#  else
+#   define DNSSD_STDCALL
+#  endif
+# else
+#  include <dlfcn.h>
+#  define DNSSD_STDCALL
+# endif
+
+typedef struct _DNSServiceRef_t *DNSServiceRef;
+#ifndef _WIN32
+typedef union _TXTRecordRef_t { char PrivateData[16]; char *ForceNaturalAlignment; } TXTRecordRef;
+#endif
+typedef uint32_t DNSServiceFlags;
+typedef int32_t  DNSServiceErrorType;
+
+typedef void (DNSSD_STDCALL *DNSServiceRegisterReply)
+    (
+    DNSServiceRef                       sdRef,
+    DNSServiceFlags                     flags,
+    DNSServiceErrorType                 errorCode,
+    const char                          *name,
+    const char                          *regtype,
+    const char                          *domain,
+    void                                *context
+    );
+
+#else
+//# include <dns_sd.h>
+# define DNSSD_STDCALL
+#endif
+
+typedef DNSServiceErrorType (DNSSD_STDCALL *DNSServiceRegister_t)
+        (
+                DNSServiceRef                       *sdRef,
+                DNSServiceFlags                     flags,
+                uint32_t                            interfaceIndex,
+                const char                          *name,
+                const char                          *regtype,
+                const char                          *domain,
+                const char                          *host,
+                uint16_t                            port,
+                uint16_t                            txtLen,
+                const void                          *txtRecord,
+                DNSServiceRegisterReply             callBack,
+                void                                *context
+        );
+typedef void (DNSSD_STDCALL *DNSServiceRefDeallocate_t)(DNSServiceRef sdRef);
+typedef void (DNSSD_STDCALL *TXTRecordCreate_t)
+        (
+                TXTRecordRef     *txtRecord,
+                uint16_t         bufferLen,
+                void             *buffer
+        );
+typedef void (DNSSD_STDCALL *TXTRecordDeallocate_t)(TXTRecordRef *txtRecord);
+typedef DNSServiceErrorType (DNSSD_STDCALL *TXTRecordSetValue_t)
+        (
+                TXTRecordRef     *txtRecord,
+                const char       *key,
+                uint8_t          valueSize,
+                const void       *value
+        );
+typedef uint16_t (DNSSD_STDCALL *TXTRecordGetLength_t)(const TXTRecordRef *txtRecord);
+typedef const void * (DNSSD_STDCALL *TXTRecordGetBytesPtr_t)(const TXTRecordRef *txtRecord);
+
+
+typedef struct dnssd_private_s {
+#ifdef WIN32
+    HMODULE module;
+#elif USE_LIBDL
+    void *module;
+#endif
+
+    DNSServiceRegister_t       DNSServiceRegister;
+    DNSServiceRefDeallocate_t  DNSServiceRefDeallocate;
+    TXTRecordCreate_t          TXTRecordCreate;
+    TXTRecordSetValue_t        TXTRecordSetValue;
+    TXTRecordGetLength_t       TXTRecordGetLength;
+    TXTRecordGetBytesPtr_t     TXTRecordGetBytesPtr;
+    TXTRecordDeallocate_t      TXTRecordDeallocate;
+
+    TXTRecordRef raop_record;
+    TXTRecordRef airplay_record;
+
+    DNSServiceRef raop_service;
+    DNSServiceRef airplay_service;
+
+} dnssd_private_t;
+
+void *
+dnssd_private_init(dnssd_t *dnssd_public, int *error)
+{
+    if (error) *error = DNSSD_ERROR_NOERROR;
+
+    dnssd_private_t *dnssd = (dnssd_private_t *) calloc(1, sizeof(dnssd_private_t));
+    if (!dnssd) {
+        if (error) *error = DNSSD_ERROR_OUTOFMEM;
+        return NULL;
+    }
+
+#ifdef WIN32
+    dnssd->module = LoadLibraryA("dnssd.dll");
+    if (!dnssd->module) {
+        if (error) *error = DNSSD_ERROR_LIBNOTFOUND;
+        free(dnssd);
+        return NULL;
+    }
+    dnssd->DNSServiceRegister = (DNSServiceRegister_t)GetProcAddress(dnssd->module, "DNSServiceRegister");
+    dnssd->DNSServiceRefDeallocate = (DNSServiceRefDeallocate_t)GetProcAddress(dnssd->module, "DNSServiceRefDeallocate");
+    dnssd->TXTRecordCreate = (TXTRecordCreate_t)GetProcAddress(dnssd->module, "TXTRecordCreate");
+    dnssd->TXTRecordSetValue = (TXTRecordSetValue_t)GetProcAddress(dnssd->module, "TXTRecordSetValue");
+    dnssd->TXTRecordGetLength = (TXTRecordGetLength_t)GetProcAddress(dnssd->module, "TXTRecordGetLength");
+    dnssd->TXTRecordGetBytesPtr = (TXTRecordGetBytesPtr_t)GetProcAddress(dnssd->module, "TXTRecordGetBytesPtr");
+    dnssd->TXTRecordDeallocate = (TXTRecordDeallocate_t)GetProcAddress(dnssd->module, "TXTRecordDeallocate");
+
+    if (!dnssd->DNSServiceRegister || !dnssd->DNSServiceRefDeallocate || !dnssd->TXTRecordCreate ||
+        !dnssd->TXTRecordSetValue || !dnssd->TXTRecordGetLength || !dnssd->TXTRecordGetBytesPtr ||
+        !dnssd->TXTRecordDeallocate) {
+        if (error) *error = DNSSD_ERROR_PROCNOTFOUND;
+        FreeLibrary(dnssd->module);
+        free(dnssd);
+        return NULL;
+    }
+#elif USE_LIBDL
+    dnssd->module = dlopen("libdns_sd.so", RTLD_LAZY);
+    if (!dnssd->module) {
+      if (error) *error = DNSSD_ERROR_LIBNOTFOUND;
+      free(dnssd);
+      return NULL;
+    }
+    dnssd->DNSServiceRegister = (DNSServiceRegister_t)dlsym(dnssd->module, "DNSServiceRegister");
+    dnssd->DNSServiceRefDeallocate = (DNSServiceRefDeallocate_t)dlsym(dnssd->module, "DNSServiceRefDeallocate");
+    dnssd->TXTRecordCreate = (TXTRecordCreate_t)dlsym(dnssd->module, "TXTRecordCreate");
+    dnssd->TXTRecordSetValue = (TXTRecordSetValue_t)dlsym(dnssd->module, "TXTRecordSetValue");
+    dnssd->TXTRecordGetLength = (TXTRecordGetLength_t)dlsym(dnssd->module, "TXTRecordGetLength");
+    dnssd->TXTRecordGetBytesPtr = (TXTRecordGetBytesPtr_t)dlsym(dnssd->module, "TXTRecordGetBytesPtr");
+    dnssd->TXTRecordDeallocate = (TXTRecordDeallocate_t)dlsym(dnssd->module, "TXTRecordDeallocate");
+
+    if (!dnssd->DNSServiceRegister || !dnssd->DNSServiceRefDeallocate || !dnssd->TXTRecordCreate ||
+        !dnssd->TXTRecordSetValue || !dnssd->TXTRecordGetLength || !dnssd->TXTRecordGetBytesPtr ||
+        !dnssd->TXTRecordDeallocate) {
+        if (error) *error = DNSSD_ERROR_PROCNOTFOUND;
+        dlclose(dnssd->module);
+        free(dnssd);
+        return NULL;
+    }
+#else
+    dnssd->DNSServiceRegister = &DNSServiceRegister;
+    dnssd->DNSServiceRefDeallocate = &DNSServiceRefDeallocate;
+    dnssd->TXTRecordCreate = &TXTRecordCreate;
+    dnssd->TXTRecordSetValue = &TXTRecordSetValue;
+    dnssd->TXTRecordGetLength = &TXTRecordGetLength;
+    dnssd->TXTRecordGetBytesPtr = &TXTRecordGetBytesPtr;
+    dnssd->TXTRecordDeallocate = &TXTRecordDeallocate;
+#endif
+
+    return (void *) dnssd;
+}
+
+void
+dnssd_private_destroy(void *private)
+{
+    if (private) {
+        dnssd_private_t *dnssd = (dnssd_private_t *) private;
+#ifdef WIN32
+        FreeLibrary(dnssd->module);
+#elif USE_LIBDL
+        dlclose(dnssd->module);
+#endif
+        free(dnssd);
+    }
+}
+
+int
+dnssd_register_raop(dnssd_t *dnssd_public, unsigned short port)
+{
+    char servname[MAX_SERVNAME];
+
+    char features[22] = {0};
+
+    assert(dnssd_public);
+    assert(dnssd_public->dnssd_private);
+    dnssd_private_t *dnssd = (dnssd_private_t *) dnssd_public->dnssd_private;    
+
+    snprintf(features, sizeof(features), "0x%X,0x%X", dnssd_public->features1, dnssd_public->features2);
+
+    dnssd->TXTRecordCreate(&dnssd->raop_record, 0, NULL);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "ch", strlen(RAOP_CH), RAOP_CH);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "cn", strlen(RAOP_CN), RAOP_CN);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "da", strlen(RAOP_DA), RAOP_DA);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "et", strlen(RAOP_ET), RAOP_ET);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "vv", strlen(RAOP_VV), RAOP_VV);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "ft", strlen(features), features);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "am", strlen(GLOBAL_MODEL), GLOBAL_MODEL);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "md", strlen(RAOP_MD), RAOP_MD);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "rhd", strlen(RAOP_RHD), RAOP_RHD);
+    switch (dnssd_public->pin_pw) {
+    case 1:
+        /* sf bit 3 0x08  means "pin required" according to https://openairplay.github.io/airplay-spec/status_flags.html  */
+        dnssd->TXTRecordSetValue(&dnssd->raop_record, "pw", strlen("true"), "true");
+        dnssd->TXTRecordSetValue(&dnssd->raop_record, "sf", strlen("0x8c"), "0x8c");
+        break;
+    case 2:
+    case 3:
+        /* sf bit 7 0x80 means "password required" according to https://openairplay.github.io/airplay-spec/status_flags.html  */
+        dnssd->TXTRecordSetValue(&dnssd->raop_record, "pw", strlen("true"), "true");
+        dnssd->TXTRecordSetValue(&dnssd->raop_record, "sf", strlen("0x84"), "0x84");
+        break;
+    default:
+        dnssd->TXTRecordSetValue(&dnssd->raop_record, "pw", strlen("false"), "false");
+        dnssd->TXTRecordSetValue(&dnssd->raop_record, "sf", strlen(RAOP_SF), RAOP_SF);
+        break;
+    }
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "sr", strlen(RAOP_SR), RAOP_SR);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "ss", strlen(RAOP_SS), RAOP_SS);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "sv", strlen(RAOP_SV), RAOP_SV);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "tp", strlen(RAOP_TP), RAOP_TP);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "txtvers", strlen(RAOP_TXTVERS), RAOP_TXTVERS);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "vs", strlen(RAOP_VS), RAOP_VS);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "vn", strlen(RAOP_VN), RAOP_VN);
+    dnssd->TXTRecordSetValue(&dnssd->raop_record, "pk", strlen(dnssd_public->pk), dnssd_public->pk);
+
+    /* Convert hardware address to string */
+    if (utils_hwaddr_raop(servname, sizeof(servname), dnssd_public->hw_addr, dnssd_public->hw_addr_len) < 0) {
+        /* FIXME: handle better */
+        return -1;
+    }
+
+    /* Check that we have bytes for 'hw@name' format */
+    if (sizeof(servname) < strlen(servname) + 1 + dnssd_public->name_len + 1) {
+        /* FIXME: handle better */
+        return -2;
+    }
+
+    strncat(servname, "@", sizeof(servname)-strlen(servname)-1);
+    strncat(servname, dnssd_public->name, sizeof(servname)-strlen(servname)-1);
+
+    /* Register the service */
+    DNSServiceErrorType retval = dnssd->DNSServiceRegister(&dnssd->raop_service, 0, 0,
+                                                          servname, "_raop._tcp",
+                                                          NULL, NULL,
+                                                          htons(port),
+                                                          dnssd->TXTRecordGetLength(&dnssd->raop_record),
+                                                          dnssd->TXTRecordGetBytesPtr(&dnssd->raop_record),
+                                                          NULL, NULL);
+
+    return (int) retval;   /* error codes are listed in Apple's dns_sd.h */
+}
+
+int
+dnssd_register_airplay(dnssd_t *dnssd_public, unsigned short port)
+{
+    char device_id[3 * MAX_HWADDR_LEN];
+    char features[22] = {0};
+
+    assert(dnssd_public);
+    assert(dnssd_public->dnssd_private);
+    dnssd_private_t *dnssd = (dnssd_private_t *) dnssd_public->dnssd_private;    
+    
+    snprintf(features, sizeof(features), "0x%X,0x%X", dnssd_public->features1, dnssd_public->features2);
+
+    /* Convert hardware address to string */
+    if (utils_hwaddr_airplay(device_id, sizeof(device_id), dnssd_public->hw_addr, dnssd_public->hw_addr_len) < 0) {
+        /* FIXME: handle better */
+        return -1;
+    }
+
+    // flags is a string representing a 20-bit flag (up to 3 hex digits)
+    dnssd->TXTRecordCreate(&dnssd->airplay_record, 0, NULL);
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "deviceid", strlen(device_id), device_id);
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "features", strlen(features), features);
+    switch (dnssd_public->pin_pw) {
+    case 1:   // display onscreen pin
+        dnssd->TXTRecordSetValue(&dnssd->airplay_record, "pw", strlen("true"), "true");
+        dnssd->TXTRecordSetValue(&dnssd->airplay_record, "flags", 3, "0x4");
+        break;  
+    case 2:  // require password
+    case 3:
+        dnssd->TXTRecordSetValue(&dnssd->airplay_record, "pw", strlen("true"), "true");
+        dnssd->TXTRecordSetValue(&dnssd->airplay_record, "flags", 3, "0x4");
+        break;
+    default:
+        dnssd->TXTRecordSetValue(&dnssd->airplay_record, "pw", strlen("false"), "false");
+        dnssd->TXTRecordSetValue(&dnssd->airplay_record, "flags", 3, "0x4");
+        break;
+    }
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "model", strlen(GLOBAL_MODEL), GLOBAL_MODEL);
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "pk", strlen(dnssd_public->pk), dnssd_public->pk);
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "pi", strlen(AIRPLAY_PI), AIRPLAY_PI);
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "srcvers", strlen(AIRPLAY_SRCVERS), AIRPLAY_SRCVERS);
+    dnssd->TXTRecordSetValue(&dnssd->airplay_record, "vv", strlen(AIRPLAY_VV), AIRPLAY_VV);
+
+    /* Register the service */
+    DNSServiceErrorType retval = dnssd->DNSServiceRegister(&dnssd->airplay_service, 0, 0,
+                                                           dnssd_public->name, "_airplay._tcp",
+                                                           NULL, NULL,
+                                                           htons(port),
+                                                           dnssd->TXTRecordGetLength(&dnssd->airplay_record),
+                                                           dnssd->TXTRecordGetBytesPtr(&dnssd->airplay_record),
+                                                           NULL, NULL);
+
+    return (int) retval;   /* error codes are listed in Apple's dns_sd.h */
+}
+
+const char *
+dnssd_get_raop_txt(dnssd_t *dnssd_public, int *length)
+{
+    assert(dnssd_public);
+    assert(dnssd_public->dnssd_private);
+    dnssd_private_t *dnssd = (dnssd_private_t *) dnssd_public->dnssd_private;    
+    assert(length);
+
+    *length = dnssd->TXTRecordGetLength(&dnssd->raop_record);
+    return dnssd->TXTRecordGetBytesPtr(&dnssd->raop_record);
+}
+
+const char *
+dnssd_get_airplay_txt(dnssd_t *dnssd_public, int *length)
+{
+    assert(dnssd_public);
+    assert(dnssd_public->dnssd_private);
+    dnssd_private_t *dnssd = (dnssd_private_t *) dnssd_public->dnssd_private;    
+    assert(length);
+
+    *length = dnssd->TXTRecordGetLength(&dnssd->airplay_record);
+    return dnssd->TXTRecordGetBytesPtr(&dnssd->airplay_record);
+}
+
+void
+dnssd_unregister_raop(dnssd_t *dnssd_public)
+{
+    assert(dnssd_public);
+    assert(dnssd_public->dnssd_private);
+    dnssd_private_t *dnssd = (dnssd_private_t *) dnssd_public->dnssd_private;    
+
+    if (!dnssd->raop_service) {
+        return;
+    }
+
+    /* Deallocate TXT record */
+    dnssd->TXTRecordDeallocate(&dnssd->raop_record);
+
+    dnssd->DNSServiceRefDeallocate(dnssd->raop_service);
+    dnssd->raop_service = NULL;
+}
+
+void
+dnssd_unregister_airplay(dnssd_t *dnssd_public)
+{
+    assert(dnssd_public);
+    assert(dnssd_public->dnssd_private);
+    dnssd_private_t *dnssd = (dnssd_private_t *) dnssd_public->dnssd_private;    
+
+    if (!dnssd->airplay_service) {
+        return;
+    }
+
+    /* Deallocate TXT record */
+    dnssd->TXTRecordDeallocate(&dnssd->airplay_record);
+
+    dnssd->DNSServiceRefDeallocate(dnssd->airplay_service);
+    dnssd->airplay_service = NULL;
+}
